Request to API AI and DB save

+++ src/main/java/com/tcc/AReader/repositories/AnnotationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tcc.areader.models.Annotation;
 
-public interface AnnotationRepository extends JpaRepository<Annotation, Integer>{
+public interface AnnotationRepository extends JpaRepository<Annotation, Long>{
     
 }
